Add files via upload
this project contains 1 code file which is written by CodeVision AVR based on ATmega32. Also, 2 different Schematic and PCB files for the main board and power supply are published.
the project uses 12 K/J-Type Thermocouples to estimate temperature in addition to one internal LM35 sensor to estimate ambient temperature.
Also, this system can record data (Temperatures) on an SD Card with a TXT file and show them online on a 64*128 GLCD.
After taking the required samples, a relay will cut off the current of any implemented machine and also exhibit it by a beep.
the required time and number of needed data to be recorded, can by modified by the operator.